Change the seed in GetHashCode of an anonymous type
In the GetHashCode method of a anonymous type from AnonymousTypeUtils,
we had been using a hardcoded seed for computing the hash code.  But it
turns out that in a real anonymous type, the seed depends on the names
of the properties in the type - it is different for each type.

So I updated our anonymous types to base its hash seed off the
properties/fields in the type, too.